atask 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,19 +8,19 @@
 
 int main(void)
 {
-	unsigned long long int a = 1, b = 2, c;
+	unsigned long int a = 1, b = 2, c;
 	int i;
 
-	printf("%lld, %lld, ", a, b);
+	printf("%ld, %ld, ", a, b);
 
 	for (i = 3; i <= 50; i++)
 	{
 		c = a + b;
 
 		if (i == 50)
-			printf("%lld\n", c);
+			printf("%ld\n", c);
 		else
-			printf("%lld, ", c);
+			printf("%ld, ", c);
 
 		a = b;
 		b = c;
